Fix nTime calculation (#44)
nTime is coming through as a reverse buffer and is properly being
parsed, but not properly being converted to an integer. Because of this,
the checks that follow could never be triggered. This change converts
the ntime buffer to a string before running it through parseInt. In
addition, an isNaN check was added for additional safety.